# docs/openapi-mobile-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Q-App Mobile API",
    "version": "v1",
    "description": "Machine-readable mobile-facing HTTP contract for Q-App v1."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Meta": {
        "type": "object",
        "required": ["requestId"],
        "properties": {
          "requestId": { "type": "string" },
          "count": { "type": "integer" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "code", "meta"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": {},
          "meta": { "$ref": "#/components/schemas/Meta" }
        }
      },
...
      "Shop": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "city",
          "countryCode",
          "addressLine1",
          "latitude",
          "longitude",
          "serviceStationsCount",
          "queueEnabled",
          "queuePaused",
          "industryType",
          "approvalStatus",
          "distanceKm",
          "queueLength",
          "estimatedWaitMin",
          "bestJoinScore",
          "bestJoinReason",
          "isFavorite",
          "reviewSummary"
        ],
        "properties": {
          "id": { "type": "string" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "city": { "type": "string" },
          "region": { "type": ["string", "null"] },
          "countryCode": { "type": "string" },
          "addressLine1": { "type": "string" },
          "postalCode": { "type": ["string", "null"] },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "publicDescription": { "type": ["string", "null"] },
          "serviceStationsCount": { "type": "integer" },
          "queueEnabled": { "type": "boolean" },
          "queuePaused": { "type": "boolean" },
          "industryType": { "type": "string" },
          "approvalStatus": { "type": "string" },
          "distanceKm": { "type": ["number", "null"] },
          "queueLength": { "type": "integer" },
          "estimatedWaitMin": { "type": "integer" },
          "bestJoinScore": { "type": "number" },
          "bestJoinReason": { "type": "string" },
          "isFavorite": { "type": "boolean" },
          "favoriteSince": { "type": "string", "format": "date-time" },
          "reviewSummary": { "$ref": "#/components/schemas/ReviewSummary" },
          "reviews": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicReview" }
          }
        }
      },
      "ReviewSummary": {
        "type": "object",
        "required": ["averageRating", "ratingCount"],
        "properties": {
          "averageRating": { "type": ["number", "null"] },
          "ratingCount": { "type": "integer" }
        }
      },
      "PublicReview": {
        "type": "object",
        "required": ["id", "rating", "comment", "createdAt", "customerName"],
        "properties": {
          "id": { "type": "string" },
          "rating": { "type": ["integer", "null"], "minimum": 1, "maximum": 5 },
          "comment": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "customerName": { "type": "string" }
        }
      },
...
      "QueueCustomer": {
        "type": "object",
        "required": ["firstName"],
        "properties": {
          "firstName": { "type": "string" }
        }
      },
      "QueueShop": {
        "type": "object",
        "required": ["id", "slug", "name", "city", "queuePaused", "calledGracePeriodMin"],
        "properties": {
          "id": { "type": "string" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "city": { "type": "string" },
          "queuePaused": { "type": "boolean" },
          "calledGracePeriodMin": { "type": "integer" }
        }
      },
      "QueueStatus": {
        "type": "object",
        "required": [
          "trackingToken",
          "sortIndex",
          "joinedAt",
          "confirmationStatus",
          "confirmationRequestedAt",
          "confirmationRespondedAt",
          "calledAt",
          "visitStatus",
          "plannedDurationMin",
          "estimatedWaitMin",
          "actualDurationMin",
          "queueJoinedAt",
          "readyAt",
          "startedAt",
          "completedAt",
          "feedbackSubmitted",
          "customer",
          "shop"
        ],
        "properties": {
          "trackingToken": { "type": "string" },
          "sortIndex": { "type": "integer" },
          "joinedAt": { "type": "string", "format": "date-time" },
          "confirmationStatus": { "type": "string" },
          "confirmationRequestedAt": { "type": ["string", "null"], "format": "date-time" },
          "confirmationRespondedAt": { "type": ["string", "null"], "format": "date-time" },
          "calledAt": { "type": ["string", "null"], "format": "date-time" },
          "visitStatus": { "type": "string" },
          "plannedDurationMin": { "type": ["integer", "null"] },
          "estimatedWaitMin": { "type": ["integer", "null"] },
          "actualDurationMin": { "type": ["integer", "null"] },
          "queueJoinedAt": { "type": ["string", "null"], "format": "date-time" },
          "readyAt": { "type": ["string", "null"], "format": "date-time" },
          "startedAt": { "type": ["string", "null"], "format": "date-time" },
          "completedAt": { "type": ["string", "null"], "format": "date-time" },
          "feedbackSubmitted": { "type": "boolean" },
          "customer": { "$ref": "#/components/schemas/QueueCustomer" },
          "shop": { "$ref": "#/components/schemas/QueueShop" },
          "position": { "type": ["integer", "null"] },
          "queueLength": { "type": "integer" },
          "nearTurnNotifiedAt": { "type": ["string", "null"], "format": "date-time" },
          "arrivalConfirmationSentAt": { "type": ["string", "null"], "format": "date-time" },
          "responseWindowEndsAt": { "type": ["string", "null"], "format": "date-time" },
          "canRespondToArrival": { "type": "boolean" }
        }
      },
      "QueueJoinStartRequest": {
        "type": "object",
        "required": ["shopSlug", "firstName", "mobileNumber"],
        "properties": {
          "shopSlug": { "type": "string" },
          "firstName": { "type": "string" },
          "mobileNumber": { "type": "string" }
        }
      },
      "QueueJoinStartResult": {
        "type": "object",
        "required": ["challengeId", "expiresAt", "message", "deliveryMode"],
        "properties": {
          "challengeId": { "type": "string" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "message": { "type": "string" },
          "deliveryMode": { "type": "string" },
          "deliveryReason": { "type": ["string", "null"] },
          "codePreview": { "type": "string" }
        }
      },
      "QueueJoinVerifyRequest": {
        "type": "object",
        "required": ["challengeId", "code"],
        "properties": {
          "challengeId": { "type": "string" },
          "code": { "type": "string" }
        }
      },
      "QueueJoinVerifyResult": {
        "type": "object",
        "required": ["alreadyJoined", "queueStatus"],
        "properties": {
          "alreadyJoined": { "type": "boolean" },
          "queueStatus": { "$ref": "#/components/schemas/QueueStatus" }
        }
      },
      "ArrivalResponseRequest": {
        "type": "object",
        "required": ["response"],
        "properties": {
          "response": {
            "type": "string",
            "enum": ["COMING", "DECLINED"]
          }
        }
      },
      "QueueFeedbackRequest": {
        "type": "object",
        "properties": {
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment": { "type": "string", "maxLength": 1000 }
        }
      },
      "QueueLeaveRequest": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 200 }
        }
      },
      "QueueLeaveResult": {
        "type": "object",
        "required": ["trackingToken", "visitStatus", "releasedAt", "message"],
        "properties": {
          "trackingToken": { "type": "string" },
          "visitStatus": { "type": "string" },
          "releasedAt": { "type": ["string", "null"], "format": "date-time" },
          "message": { "type": "string" }
        }
      },
      "QueueFeedbackResult": {
        "type": "object",
        "required": ["feedbackSubmitted", "message"],
        "properties": {
          "feedbackSubmitted": { "type": "boolean" },
          "message": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Invalid payload.",
              "code": "BAD_REQUEST",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Resource not found.",
              "code": "NOT_FOUND",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Conflict.",
              "code": "CONFLICT",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/shops": {
      "get": {
        "tags": ["Discovery"],
        "operationId": "listShops",
        "summary": "List public live shops",
        "parameters": [
          {
            "name": "X-QApp-Device-Id",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "Anonymous app/browser device id used to mark favourite shops before customer login."
          },
          { "name": "latitude", "in": "query", "schema": { "type": "number" } },
          { "name": "longitude", "in": "query", "schema": { "type": "number" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Public shop list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Shop" }
                    },
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Shop" }
                    },
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": [
                    {
                      "id": "shop-id",
                      "slug": "demo-barber",
                      "name": "Fade Yard Soho",
                      "city": "London",
                      "region": null,
                      "countryCode": "GB",
                      "addressLine1": "24 Berwick Street",
                      "postalCode": "W1F 8RG",
                      "latitude": 51.5137,
                      "longitude": -0.1366,
                      "publicDescription": "Live demo location.",
                      "serviceStationsCount": 2,
                      "queueEnabled": true,
                      "queuePaused": false,
                      "industryType": "BARBER",
                      "approvalStatus": "APPROVED",
                      "distanceKm": 0,
                      "queueLength": 4,
                      "estimatedWaitMin": 40,
                      "bestJoinScore": 40,
                      "bestJoinReason": "Balanced by distance and live wait",
                      "isFavorite": false,
                      "reviewSummary": {
                        "averageRating": 4.8,
                        "ratingCount": 24
                      }
                    }
                  ],
                  "items": [],
                  "meta": { "count": 1, "requestId": "request-id" }
                }
              }
            }
          }
        }
      }
    },
    "/shops/{slug}": {
      "get": {
        "tags": ["Discovery"],
        "operationId": "getShopBySlug",
        "summary": "Get one public shop by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Public shop detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Shop" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "id": "shop-id",
                    "slug": "demo-barber",
                    "name": "Fade Yard Soho",
                    "city": "London",
                    "region": null,
                    "countryCode": "GB",
                    "addressLine1": "24 Berwick Street",
                    "postalCode": "W1F 8RG",
                    "latitude": 51.5137,
                    "longitude": -0.1366,
                    "publicDescription": "Live demo location.",
                    "serviceStationsCount": 2,
                    "queueEnabled": true,
                    "queuePaused": false,
                    "industryType": "BARBER",
                    "approvalStatus": "APPROVED",
                    "distanceKm": null,
                    "queueLength": 4,
                    "estimatedWaitMin": 40,
                    "bestJoinScore": 52,
                    "bestJoinReason": "Selected shop",
                    "isFavorite": false,
                    "reviewSummary": {
                      "averageRating": 4.8,
                      "ratingCount": 24
                    },
                    "reviews": [
                      {
                        "id": "review-id-1",
                        "rating": 5,
                        "comment": "Got the leave-now alert at exactly the right time.",
                        "createdAt": "2026-06-12T10:15:00.000Z",
                        "customerName": "Amir"
                      },
                      {
                        "id": "review-id-2",
                        "rating": 4,
                        "comment": "Easy queue, no standing around outside.",
                        "createdAt": "2026-06-11T16:30:00.000Z",
                        "customerName": "Sofia"
                      }
                    ]
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/queue/join/start": {
      "post": {
        "tags": ["Queue"],
        "operationId": "startQueueJoin",
        "summary": "Start queue join with OTP challenge",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueueJoinStartRequest" },
              "example": {
                "shopSlug": "demo-barber",
                "firstName": "Amir",
                "mobileNumber": "+447400000301"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP challenge created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/QueueJoinStartResult" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "challengeId": "challenge-id",
                    "expiresAt": "2026-06-12T09:30:00.000Z",
                    "message": "Verification code created. SMS is currently using preview mode.",
                    "deliveryMode": "preview",
                    "deliveryReason": "twilio_not_configured"
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/queue/join/verify": {
      "post": {
        "tags": ["Queue"],
        "operationId": "verifyQueueJoin",
        "summary": "Verify OTP and join queue",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueueJoinVerifyRequest" },
              "example": {
                "challengeId": "challenge-id",
                "code": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queue join verified",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/QueueJoinVerifyResult" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "alreadyJoined": false,
                    "queueStatus": {
                      "trackingToken": "demo-queue-amir",
                      "sortIndex": 1,
                      "joinedAt": "2026-06-12T09:21:00.000Z",
                      "confirmationStatus": "PENDING",
                      "confirmationRequestedAt": null,
                      "confirmationRespondedAt": null,
                      "calledAt": null,
                      "visitStatus": "QUEUED",
                      "plannedDurationMin": 20,
                      "estimatedWaitMin": 18,
                      "actualDurationMin": null,
                      "queueJoinedAt": "2026-06-12T09:21:00.000Z",
                      "readyAt": null,
                      "startedAt": null,
                      "completedAt": null,
                      "feedbackSubmitted": false,
                      "customer": {
                        "firstName": "Amir"
                      },
                      "shop": {
                        "id": "shop-id",
                        "slug": "demo-barber",
                        "name": "Fade Yard Soho",
                        "city": "London",
                        "queuePaused": false,
                        "calledGracePeriodMin": 5
                      },
                      "position": 1,
                      "queueLength": 4,
                      "nearTurnNotifiedAt": null,
                      "arrivalConfirmationSentAt": null,
                      "responseWindowEndsAt": null,
                      "canRespondToArrival": false
                    }
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/queue/status/{trackingToken}": {
      "get": {
        "tags": ["Queue"],
        "operationId": "getQueueStatus",
        "summary": "Get queue status by tracking token",
        "parameters": [
          {
            "name": "trackingToken",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Queue status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/QueueStatus" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "trackingToken": "demo-queue-amir",
                    "sortIndex": 1,
                    "joinedAt": "2026-06-12T09:21:00.000Z",
                    "confirmationStatus": "PENDING",
                    "confirmationRequestedAt": "2026-06-12T09:32:00.000Z",
                    "confirmationRespondedAt": null,
                    "calledAt": null,
                    "visitStatus": "CONFIRMATION_PENDING",
                    "plannedDurationMin": 20,
                    "estimatedWaitMin": 8,
                    "actualDurationMin": null,
                    "queueJoinedAt": "2026-06-12T09:21:00.000Z",
                    "readyAt": null,
                    "startedAt": null,
                    "completedAt": null,
                    "feedbackSubmitted": false,
                    "customer": {
                      "firstName": "Amir"
                    },
                    "shop": {
                      "id": "shop-id",
                      "slug": "demo-barber",
                      "name": "Fade Yard Soho",
                      "city": "London",
                      "queuePaused": false,
                      "calledGracePeriodMin": 5
                    },
                    "position": 1,
                    "queueLength": 2,
                    "nearTurnNotifiedAt": "2026-06-12T09:31:30.000Z",
                    "arrivalConfirmationSentAt": "2026-06-12T09:32:00.000Z",
                    "responseWindowEndsAt": "2026-06-12T09:37:00.000Z",
                    "canRespondToArrival": true
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/queue/status/{trackingToken}/respond-arrival": {
      "post": {
        "tags": ["Queue"],
        "operationId": "respondArrival",
        "summary": "Respond to arrival confirmation",
        "parameters": [
          {
            "name": "trackingToken",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArrivalResponseRequest" },
              "example": {
                "response": "COMING"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Arrival response recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": ["trackingToken", "response"],
                      "properties": {
                        "trackingToken": { "type": "string" },
                        "response": { "type": "string" }
                      }
                    },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "trackingToken": "demo-queue-amir",
                    "response": "COMING"
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/queue/status/{trackingToken}/leave": {
      "post": {
        "tags": ["Queue"],
        "operationId": "leaveQueue",
        "summary": "Leave an active queue place from the secure status link",
        "parameters": [
          {
            "name": "trackingToken",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueueLeaveRequest" },
              "example": {
                "reason": "Customer left queue from status page"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queue place released",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/QueueLeaveResult" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "trackingToken": "demo-queue-amir",
                    "visitStatus": "CANCELLED",
                    "releasedAt": "2026-06-12T09:29:00.000Z",
                    "message": "Your queue place has been released."
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/queue/status/{trackingToken}/feedback": {
      "post": {
        "tags": ["Queue"],
        "operationId": "submitQueueFeedback",
        "summary": "Submit optional feedback after a completed queue visit",
        "parameters": [
          {
            "name": "trackingToken",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueueFeedbackRequest" },
              "example": {
                "rating": 5,
                "comment": "Easy to join and I did not have to wait inside."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/QueueFeedbackResult" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "feedbackSubmitted": true,
                    "message": "Thanks for helping this shop improve."
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  }
}